pyproject.toml
[tool.poetry]
name = "githubctl"
version = "0.1.0"
description = ""
authors = ["Peng Xiao <xiaoquwl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
